+++ tests/after_hook.rs
@@ -1,4 +1,5 @@
 use std::{
+    future,
     panic::AssertUnwindSafe,
     sync::atomic::{AtomicUsize, Ordering},
     time::Duration,
@@ -11,6 +12,10 @@ use tokio::time;
 
 static NUMBER_OF_BEFORE_WORLDS: AtomicUsize = AtomicUsize::new(0);
 static NUMBER_OF_AFTER_WORLDS: AtomicUsize = AtomicUsize::new(0);
+static NUMBER_OF_FAILED_HOOKS: AtomicUsize = AtomicUsize::new(0);
+static NUMBER_OF_PASSED_STEPS: AtomicUsize = AtomicUsize::new(0);
+static NUMBER_OF_SKIPPED_STEPS: AtomicUsize = AtomicUsize::new(0);
+static NUMBER_OF_FAILED_STEPS: AtomicUsize = AtomicUsize::new(0);
 
 #[tokio::main]
 async fn main() {
@@ -23,17 +28,28 @@ async fn main() {
             }
             .boxed()
         })
-        .after(move |_, _, _, w| {
-            async move {
-                if w.is_some() {
-                    let after =
-                        NUMBER_OF_AFTER_WORLDS.fetch_add(1, Ordering::SeqCst);
-                    assert_ne!(after, 8, "Too much after `World`s!");
-                } else {
-                    panic!("No World received");
-                }
+        .after(move |_, _, _, ev, w| {
+            use cucumber::event::ScenarioFinished::{
+                BeforeHookFailed, StepFailed, StepPassed, StepSkipped,
+            };
+
+            match ev {
+                BeforeHookFailed(_) => &NUMBER_OF_FAILED_HOOKS,
+                StepPassed => &NUMBER_OF_PASSED_STEPS,
+                StepSkipped => &NUMBER_OF_SKIPPED_STEPS,
+                StepFailed(_, _, _) => &NUMBER_OF_FAILED_STEPS,
             }
-            .boxed()
+            .fetch_add(1, Ordering::SeqCst);
+
+            if w.is_some() {
+                let after =
+                    NUMBER_OF_AFTER_WORLDS.fetch_add(1, Ordering::SeqCst);
+                assert_ne!(after, 8, "Too much after `World`s!");
+            } else {
+                panic!("No `World` received");
+            }
+
+            future::ready(()).boxed()
         })
         .run_and_exit("tests/features/wait");
 
@@ -46,6 +62,10 @@ async fn main() {
     assert_eq!(err, "2 steps failed, 1 parsing error, 4 hook errors");
     assert_eq!(NUMBER_OF_BEFORE_WORLDS.load(Ordering::SeqCst), 11);
     assert_eq!(NUMBER_OF_AFTER_WORLDS.load(Ordering::SeqCst), 11);
+    assert_eq!(NUMBER_OF_FAILED_HOOKS.load(Ordering::SeqCst), 2);
+    assert_eq!(NUMBER_OF_PASSED_STEPS.load(Ordering::SeqCst), 6);
+    assert_eq!(NUMBER_OF_SKIPPED_STEPS.load(Ordering::SeqCst), 2);
+    assert_eq!(NUMBER_OF_FAILED_STEPS.load(Ordering::SeqCst), 2);
 }
 
 #[given(regex = r"(\d+) secs?")]